refactor : Renamed package winningLotto->winning
