+++ src/main/java/domain/lotto/Lottos.java
@@ -1,7 +1,7 @@
 package domain.lotto;
 
 import domain.statistics.WinningRank;
-import domain.winningLotto.WinningLotto;
+import domain.winning.WinningLotto;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
