+++ src/test/java/domain/winning/WinningLottoTest.java
@@ -1,4 +1,4 @@
-package domain.winningLotto;
+package domain.winning;
 
 import static org.assertj.core.api.Assertions.assertThat;
